+++ lib/facter/security_baseline/common/check_package_installed.rb
@@ -10,7 +10,11 @@ def check_package_installed(pkg, opts = '-q')
   if (os == 'RedHat') || (os == 'Suse')
     val = Facter::Core::Execution.exec("rpm #{opts} #{pkg}")
   elsif os == 'Debian'
-    val = Facter::Core::Execution.exec("dpkg #{opts} #{pkg}")
+    if opts == '-q'
+      opts = ''
+    end
+    opts = "#{opts} -l"
+    val = Facter::Core::Execution.exec("dpkg #{opts} #{pkg} | grep ^ii")
   end
   if val.nil? || val.empty? || val =~ %r{not installed}
     false