+++ spec/classes/rules/redhat/sec_crypto_policy_spec.rb
@@ -31,7 +31,7 @@ describe 'security_baseline::rules::redhat::sec_crypto_policy' do
       it {
         is_expected.to compile
         if enforce
-          is_expected.to contain_exec('set crypto policy to FUTURE')
+          is_expected.to contain_exec('set crypto policy to FUTURE (current: DEFAULT)')
             .with(
               'command' => 'update-crypto-policies --set FUTURE',
               'path'    => ['/sbin', '/usr/sbin', '/bin', '/usr/bin'],
@@ -45,7 +45,7 @@ describe 'security_baseline::rules::redhat::sec_crypto_policy' do
 
           is_expected.not_to contain_echo('crypto-policy')
         else
-          is_expected.not_to contain_exec('set crypto policy to FUTURE')
+          is_expected.not_to contain_exec('set crypto policy to FUTURE (current: DEFAULT)')
           is_expected.not_to contain_exec('set FIPS to disable')
           is_expected.to contain_echo('crypto-policy')
             .with(
@@ -84,7 +84,7 @@ describe 'security_baseline::rules::redhat::sec_crypto_policy' do
       it {
         is_expected.to compile
         if enforce
-          is_expected.to contain_exec('set crypto policy to FIPS')
+          is_expected.to contain_exec('set crypto policy to FIPS (current: DEFAULT)')
             .with(
               'command' => 'update-crypto-policies --set FIPS',
               'path'    => ['/sbin', '/usr/sbin', '/bin', '/usr/bin'],
@@ -98,7 +98,7 @@ describe 'security_baseline::rules::redhat::sec_crypto_policy' do
 
           is_expected.not_to contain_echo('crypto-policy')
         else
-          is_expected.not_to contain_exec('set crypto policy to FIPS')
+          is_expected.not_to contain_exec('set crypto policy to FIPS (current: DEFAULT)')
           is_expected.not_to contain_exec('set FIPSto enable')
           is_expected.to contain_echo('crypto-policy')
             .with(